Nouveau fichier pour segmenter le code
